Refactor resource classes to be in separate files
This commit refactors resource based classes to be in separate files.
